# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ViaX Scout SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  birthDate?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  birthDate?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface UpdateProfileBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  birthDate?: string | null;
}

export interface UpdatePasswordBody {
  currentPassword: string;
  newPassword: string;
}

/**
 * builtin = embedded parser, ai = AI-powered
 */
export type UserSettingsParserMode =
  (typeof UserSettingsParserMode)[keyof typeof UserSettingsParserMode];

export const UserSettingsParserMode = {
  builtin: "builtin",
  ai: "ai",
} as const;

/**
 * Geocoding instance: builtin=Nominatim/OSM, googlemaps=Google Maps API
 */
export type UserSettingsInstanceMode =
  (typeof UserSettingsInstanceMode)[keyof typeof UserSettingsInstanceMode];

export const UserSettingsInstanceMode = {
  builtin: "builtin",
  googlemaps: "googlemaps",
} as const;

export interface UserSettings {
  /** builtin = embedded parser, ai = AI-powered */
  parserMode: UserSettingsParserMode;
  /** @nullable */
  aiProvider?: string | null;
  /** @nullable */
  aiApiKey?: string | null;
  toleranceMeters: number;
  /** Geocoding instance: builtin=Nominatim/OSM, googlemaps=Google Maps API */
  instanceMode: UserSettingsInstanceMode;
  /** @nullable */
  googleMapsApiKey?: string | null;
}

export type UpdateSettingsBodyParserMode =
  (typeof UpdateSettingsBodyParserMode)[keyof typeof UpdateSettingsBodyParserMode];

export const UpdateSettingsBodyParserMode = {
  builtin: "builtin",
  ai: "ai",
} as const;

export type UpdateSettingsBodyInstanceMode =
  (typeof UpdateSettingsBodyInstanceMode)[keyof typeof UpdateSettingsBodyInstanceMode];

export const UpdateSettingsBodyInstanceMode = {
  builtin: "builtin",
  googlemaps: "googlemaps",
} as const;

export interface UpdateSettingsBody {
  parserMode?: UpdateSettingsBodyParserMode;
  /** @nullable */
  aiProvider?: string | null;
  /** @nullable */
  aiApiKey?: string | null;
  toleranceMeters?: number;
  instanceMode?: UpdateSettingsBodyInstanceMode;
  /** @nullable */
  googleMapsApiKey?: string | null;
}

export type AnalysisStatus =
  (typeof AnalysisStatus)[keyof typeof AnalysisStatus];

export const AnalysisStatus = {
  pending: "pending",
  processing: "processing",
  done: "done",
  error: "error",
} as const;

export interface Analysis {
  id: number;
  userId: number;
  fileName: string;
  totalAddresses: number;
  nuances: number;
  geocodeSuccess: number;
  similarityAvg: number;
  processingTimeMs: number;
  parserMode: string;
  status: AnalysisStatus;
  /**
   * JSON string of detailed results
   * @nullable
   */
  results?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AnalysisSummary {
  id: number;
  fileName: string;
  totalAddresses: number;
  nuances: number;
  status: string;
  createdAt: string;
}

export interface AnalysesList {
  items: Analysis[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateAnalysisBody {
  fileName: string;
  totalAddresses: number;
  nuances: number;
  geocodeSuccess: number;
  similarityAvg: number;
  processingTimeMs: number;
  parserMode: string;
  /** @nullable */
  results?: string | null;
}

export interface DashboardSummary {
  totalAnalyses: number;
  totalAddressesProcessed: number;
  avgNuanceRate: number;
  avgGeocodeSuccess: number;
  avgSimilarity: number;
  analysesThisMonth: number;
}

export type UploadAvatarBody = {
  avatar: Blob;
};

export type ListAnalysesParams = {
  page?: number;
  limit?: number;
};

export type ProcessUploadBody = {
  arquivo: Blob;
};