# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ViaX Scout SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload profile avatar from gallery
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required:
                - avatar
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/password:
    patch:
      operationId: updatePassword
      tags: [users]
      summary: Update password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordBody"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/settings:
    get:
      operationId: getSettings
      tags: [users]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      operationId: updateSettings
      tags: [users]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
    UpdatePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    UserSettings:
      type: object
      properties:
        parserMode:
          type: string
          enum: [builtin, ai]
          description: "builtin = embedded parser, ai = AI-powered"
        aiProvider:
          type: ["string", "null"]
        aiApiKey:
          type: ["string", "null"]
        toleranceMeters:
          type: integer
        instanceMode:
          type: string
          enum: [builtin, googlemaps]
          description: "Geocoding instance: builtin=Nominatim/OSM, googlemaps=Google Maps API"
        googleMapsApiKey:
          type: ["string", "null"]
      required:
        - parserMode
        - toleranceMeters
        - instanceMode
    UpdateSettingsBody:
      type: object
      properties:
        parserMode:
          type: string
          enum: [builtin, ai]
        aiProvider:
          type: ["string", "null"]
        aiApiKey:
          type: ["string", "null"]
        toleranceMeters:
          type: integer
        instanceMode:
          type: string
          enum: [builtin, googlemaps]
        googleMapsApiKey:
          type: ["string", "null"]
...